moved AbstractConfigBuilderBasedLaunchConfigurationDelegate ti de.uka.ipd.sdq.workflow.launchconfig.core
