+++ bundles/de.uka.ipd.sdq.workflow.launchconfig.core/META-INF/MANIFEST.MF
@@ -6,7 +6,8 @@ Bundle-Version: 5.2.1.qualifier
 Export-Package: de.uka.ipd.sdq.workflow.launchconfig.core,
  de.uka.ipd.sdq.workflow.launchconfig.core.configbuilder
 Require-Bundle: org.eclipse.debug.core;bundle-version="3.20.0",
- de.uka.ipd.sdq.workflow;bundle-version="5.2.1"
+ de.uka.ipd.sdq.workflow;bundle-version="5.2.1",
+ de.uka.ipd.sdq.workflow.logging.console;bundle-version="5.2.1"
 Bundle-ActivationPolicy: lazy
 Bundle-Vendor: sdq.ipd.uka.de
 Bundle-RequiredExecutionEnvironment: JavaSE-17
