+++ bundles/de.uka.ipd.sdq.workflow.launchconfig/src/de/uka/ipd/sdq/workflow/launchconfig/tabs/DebugEnabledCommonTab.java
@@ -15,7 +15,7 @@ import org.eclipse.swt.widgets.Control;
 import org.eclipse.swt.widgets.Group;
 import org.eclipse.swt.widgets.Label;
 
-import de.uka.ipd.sdq.workflow.launchconfig.AbstractWorkflowBasedLaunchConfigurationDelegate;
+import de.uka.ipd.sdq.workflow.launchconfig.core.AbstractWorkflowBasedLaunchConfigurationDelegate;
 
 /**
  * The Class DebugEnabledCommonTab.
